removing old debug file
